backend/pyproject.toml
[project]
name = "doqlink-backend"
version = "0.1.0"
description = "DoqLink AI backend"
authors = [
    {name = "nFactorial-AI-Cup-2025"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "openai (>=1.82.0,<2.0.0)",
    "playwright (>=1.52.0,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "fastapi-users[sqlalchemy] (>=14.0.1,<15.0.0)",
    "fastapi-pagination (>=0.13.1,<0.14.0)",
    "fastapi-cache2 (>=0.2.2,<0.3.0)",
    "fastapi-limiter (>=0.1.6,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
